Unsuccessful attempt to improve contour coloring
I attempted to filter the contours using a smaller bounding box that excludes the margins of the viewport.
This didn't improve things much in the Mount Diablo Summit case, because the real problem is the lower-elevation contours with are in the viewport but fully occluded by the higher elevation contours. I'm still not aware of any good way to filter out the fully occluded contours.